requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
slowapi
celery[redis]
redis
python-dotenv
